Revert "possible to use simplify for each type no just list "
