+++ requre/storage.py
@@ -381,9 +381,6 @@ class PersistentObjectStorage(metaclass=SingletonMeta):
             self.metadata = {
                 self.key_inspect_strategy_key: DataMiner().key_stategy_cls.__name__
             }
-        data_type_name = DataMiner().data_type.__class__.__name__
-        if self.metadata.get(data_type_name) is None:
-            self.metadata = {data_type_name: DataMiner().data_type.value}
 
     @property
     def storage_file(self):
@@ -459,12 +456,13 @@ class PersistentObjectStorage(metaclass=SingletonMeta):
         debug_keys: List[str] = []
         matched_calls: List[str] = []
         list_len = len(hashable_keys)
-        for index, item in enumerate(hashable_keys):
+        for item_num in range(list_len):
+            item = hashable_keys[item_num]
             if item not in current_level:
                 # it matched last 2 items
                 if DataMiner().read_key_exact or (
                     not DataMiner().read_key_exact
-                    and index + KEY_MINIMAL_MATCH >= list_len
+                    and item_num + KEY_MINIMAL_MATCH >= list_len
                 ):
                     # if not matched, but consider if it is not same key as previous
                     # it is important if simplify used.
@@ -474,7 +472,7 @@ class PersistentObjectStorage(metaclass=SingletonMeta):
                     raise PersistentStorageException(
                         f"Keys not in storage:{self.storage_file}"
                         f" Matched: {debug_keys},"
-                        f" Missing: {hashable_keys[index:]}"
+                        f" Missing: {hashable_keys[item_num:]}"
                     )
                 else:
                     debug_keys.append(f"SKIP {item}")
